Connect to step on all allocated nodes concurrently
For each node that the step is running on, we create a `Connection`
instance.  Each connection creates an "input" pipe and `STDIN` is now
written linewise to the pipe for each connection.  Each connection
continues to stream its output to `STDOUT`.

This gets us the same behaviour as slurms `srun` command when the
`--pty` flag is not given.

Obviously, this is not going to be sufficient to support PTY sessions,
but they really do need special handling.